FT-03C sortby function changed
