+++ app/components/users-cards.js
@@ -4,13 +4,28 @@ import { computed, get, set } from '@ember/object';
 
 export default Component.extend({
   store: service(),
+  init(){
+    this._super(...arguments);
+    set(this,'sortName','firstName');
+    set(this, 'selectedOrder', 'asc');
+    set(this, 'selectedValue', 'firstName');
+
+  },
 
   allUsers: computed(function () {
-    return get(this,'store').findAll('user', { reload: true });
+    this.getUser = get(this,'store').findAll('user', { reload: true });
+    return this.getUser;
   }),
 
-  userList: computed('getTeam', function () {
-    return this.getTeam ? this.allUsers.filterBy('team', this.getTeam) : this.allUsers;
+  userList: computed('getTeam','sortName','selectedOrder','allUsers.isFulfilled', function () {
+    if (this.allUsers.isFulfilled) {
+      if (this.selectedOrder == 'desc' ) {
+        return  this.getTeam ? this.allUsers.filterBy('team', this.getTeam).sortBy(this.sortName).reverse() : this.allUsers.sortBy(this.sortName).reverse();
+    } else{
+      return  this.getTeam ? this.allUsers.filterBy('team', this.getTeam).sortBy(this.sortName) : this.allUsers.sortBy(this.sortName);
+    }
+  }
+
   }),
 
   myTeamMembers: computed('allUsers.isFulfilled', function () {
@@ -19,14 +34,68 @@ export default Component.extend({
     }
   }),
 
+
+
+  sortArray: computed(function () {
+    let sortList = [
+      {
+        name: 'firstName',
+        displayName: 'First Name',
+      },
+      {
+        name: 'lastName',
+        displayName: 'Last Name',
+      },
+      {
+        name: 'joiningDate',
+        displayName: 'Joining Date',
+      },
+    ];
+
+    return sortList;
+  }),
+
+  sortOrderArray: computed(function () {
+    let sortOrderList = [
+      {
+        orderParam: 'asc',
+        sortOrder: 'Ascending',
+      },
+      {
+        orderParam: 'desc',
+        sortOrder: 'Descending',
+      },
+    ];
+
+    return sortOrderList;
+  }),
+
+
+
   actions: {
-    touchEnd: function () {
-      this.toggleProperty('isShowingBody');
+    filterby: function () {
       this.toggleProperty('clicked');
+    },
+
+    sortby: function () {
+      this.toggleProperty('sortByclicked');
     },
 
     isFilter: async function (team) {
       set(this,'getTeam', team);
+      set(this, 'currentTeam', team);
     },
+
+    sortFunc: function (sortName) {
+        set(this, 'sortName', sortName);
+        set(this, 'selectedValue', sortName);
+
+    },
+
+    sortOrder: function (order) {
+       set(this, 'selectedOrder', order);
+    },
+
+
   },
 });
