+++ app/styles/sub-styles/filter-nav.scss
@@ -11,7 +11,6 @@
 
 .filter-by::after{
   content: none;
-
 }
 
 .filter-by:hover{
@@ -56,12 +55,10 @@
 
 
 .filter .team {
-  padding-top: 5px;
   font-size: 15px;
   font-weight: 500;
-  cursor: default;
-  position: relative;
-  right: 12px;
+  margin-top: 10px;
+  cursor: pointer;
 }
 
 .filter-table {
@@ -71,6 +68,7 @@
   color: #12344d;
   opacity: .8;
 }
+
 
 .filter-table button{
   border: none;
