--- app/templates/components/filter-by.hbs
@@ -1,60 +1,0 @@
-{{! Filter By }}
-
-<div class="filter">
-  <BsDropdown as |dd|>
-    <dd.toggle class="filter-by"><span>Filter by</span><p
-        {{action "filterby"}}
-      >{{if currentTeam currentTeam "All Employees"}}</p>
-      <img
-        src="/assets/images/accordion-open.svg"
-        alt="img"
-        class="open {{if clicked 'turn'}}"
-      />
-    </dd.toggle>
-    <dd.menu>
-      <li class="team">Teams</li>
-      {{#each myTeamMembers as |myTeam|}}
-        <li class="filter-table"><button
-            {{action "isFilter" myTeam.team}}
-          >{{myTeam.team}}</button></li>
-      {{/each}}
-    </dd.menu>
-  </BsDropdown>
-</div>
-
-{{! Sort By }}
-
-<div class="sort">
-  <BsDropdown as |dd|>
-    <dd.toggle class="filter-by"><span>Sort by</span><p {{action "sortby"}}>{{if
-          sortDisplay
-          sortDisplay
-          "First Name"
-        }}</p>
-      <img
-        src="/assets/images/accordion-open.svg"
-        alt="img"
-        class="open {{if sortByclicked 'turn'}}"
-      />
-    </dd.toggle>
-    <dd.menu>
-      <div class="menu">
-        {{#each sortArray as |sortType|}}
-          <li><button
-              class="sort-table
-                {{if (eq this.selectedValue sortType.name) 'tick'}}"
-              {{action "sortFunc" sortType.name sortType.displayName}}
-            >{{sortType.displayName}}</button></li>
-        {{/each}}
-        <div class="dropdown-divider divider"></div>
-        {{#each sortOrderArray as |sortOrderType|}}
-          <li><button
-              class="sort-method
-                {{if (eq selectedOrder sortOrderType.orderParam) 'tick'}}"
-              {{action "sortOrder" sortOrderType.orderParam}}
-            >{{sortOrderType.sortOrder}}</button></li>
-        {{/each}}
-      </div>
-    </dd.menu>
-  </BsDropdown>
-</div>
