--- app/templates/components/user-count.hbs
@@ -1,7 +1,0 @@
-{{#filter-by usersList=getUser}}{{/filter-by}}
-
-<div class="userWrapper">
-  {{#each this.getUser as |user|}}
-    {{#user-card user=user}}{{/user-card}}
-  {{/each}}
-</div>
